requirements.txt
dearpygui~=1.10.1
pytest~=7.4.4
pyserial~=3.5
tomli~=2.0.1
tomli_w~=1.0.0
future~=0.18.3
crc~=6.1.0
bidict~=0.22.1
pymavlink~=2.4.41